Revert "[Fix] Change 'long int' to 'unsigned int'"
BUG found in:
  FILE: random.c
  FUNCTION: random_in_range_exclusive

  Overflow in arithmetic multiplication operation.

The rand() function from the <stdlib.h> library generates a number in
the range [0, RAND_MAX]. In the GNU library, RAND_MAX is 2147483647,
which is the largest signed integer representable in 32 bits. The issue
raises when executing: rand() * length; being 'length' declared as
'unsigned int'. GCC compiles gives an overflow warning.

According to the C standard ISO/IEC 9899:201x (2011) Section 6.3.1.1-2,
both, the number generated by 'rand()' and 'length' are promoted to
'int' since it can represent all their possible values.
This, leads to GCC to warn us because the operation might cause an
overflow which is considered 'undefined behaviour'.

Additional reference:
http://stackoverflow.com/questions/15656437/int-promotion-is-the-following-well-defined

This reverts commit 9967e88a1bb5ee6e455d3a51ec2ddd47619bb30f.

+++ src/genome.c
@@ -34,8 +34,8 @@
     Encoding.
 */
 
-struct Encoding *
-create_encoding(unsigned int *nucleotides,
+struct Encoding * 
+create_encoding(long int *nucleotides,
                 int nucleotides_length,
                 int genes_length)
 {
@@ -47,7 +47,7 @@ create_encoding(unsigned int *nucleotides,
     return e;
 }
 
-unsigned int
+long int
 min_nucleotide_value(struct Encoding *encoding)
 {
     if (!encoding) {
@@ -55,7 +55,7 @@ min_nucleotide_value(struct Encoding *encoding)
     }
     return encoding->nucleotides[0];
 }
-unsigned int
+long int 
 max_nucleotide_value(struct Encoding *encoding)
 {
     if (!encoding) {
@@ -73,7 +73,7 @@ create_individual(struct Encoding *encoding)
 {
     struct Individual *new = (struct Individual *)
                                 malloc(sizeof(struct Individual));
-    unsigned int *genes = (unsigned int *) malloc(encoding->genes_length * GENE_BYTES);
+    long int *genes = (long int *) malloc(encoding->genes_length * GENE_BYTES);
 
     memset(genes, 0, encoding->genes_length * GENE_BYTES);
     new->genes = genes;
@@ -127,7 +127,7 @@ create_empty_population(struct Encoding *encoding,
 struct Population *
 create_random_population(struct Encoding *encoding,
                             int initial_size,
-                            int max_size)
+                            int max_size)  
 {
     struct Population *population = create_empty_population(encoding, max_size);
     struct Individual *new;
@@ -140,7 +140,7 @@ create_random_population(struct Encoding *encoding,
     return population;
 }
 
-int
+int 
 add_individual(struct Population *population,
                 struct Individual *new)
 {
